+++ grunt/config/htmlmin.js
@@ -1,0 +1,27 @@
+/**
+ * Generated header by C0ZEN for x-days-later project
+ * Generated file htmlmin on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/01/2018
+ * Time: 21:48
+ * Version: 0.12.8
+ */
+module.exports = {
+	defaultTemplate: {
+		options: {
+			minifyCSS            : true,
+			minifyJS             : true,
+			removeComments       : true,
+			useShortDoctype      : true,
+			sortAttributes       : true,
+			sortClassName        : true,
+			decodeEntities       : true,
+			collapseWhitespace   : true,
+			removeAttributeQuotes: true
+		},
+		files  : {
+			'docs/_layouts/default.html': 'docs/_layouts/default.clean.html'
+		}
+	}
+};