+++ grunt/config/sass.js
@@ -1,0 +1,21 @@
+/**
+ * Generated header by C0ZEN for x-days-later project
+ * Generated file sass on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/01/2018
+ * Time: 21:41
+ * Version: 0.12.8
+ */
+module.exports = {
+	loader: {
+		options: {
+			outputStyle   : 'compressed',
+			precision     : 2,
+			sourceComments: false
+		},
+		files  : {
+			'docs/assets/css/loader.css': 'docs/assets/css/loader.scss'
+		}
+	}
+};